Add `ClientBuilder` that lets users add custom middleware without full stack replacement
